# [SPARK-27436][SQL] Add spark.sql.optimizer.nonExcludedRules