@gatorsmile 2019-04-11
The semantics of this conf is confusing. We have a built-in list of non-excludable rules. This PR tries to add a user configurable list. Anybody is requesting this feature?