@dongjoon-hyun 2019-04-11
Thank you for review, @MaxGekk and @gatorsmile .

@gatorsmile The built-in list is used to protect only Apache Spark optimizer rules. And, here, the problem is that we cannot change it at all without building from the sources (after changing the built-in list or updating a custom `Optimizer`) and distributing to the cluster.

We know `set spark.sql.optimizer.excludedRules` is convenient for the dev environments for debugging and also a good safe-guard to remove malfunctioning optimizers in the production environments. However, in the same way, this is too dangerous for the production environments because all user can do this and make a damage to the whole executions mistakenly.

We gave a too powerful sword, `set spark.sql.optimizer.excludedRules`, to the production users. Technically, we need to provide a shield, too.

BTW, at the beginning of `spark.sql.optimizer.excludedRules`, I already suggested to make `spark.sql.optimizer.excludedRules` as a static configuration, but it's not accepted because the dynamic debugging was preferred by the author of the PR.