@HyukjinKwon 2019-04-11
However, @dongjoon-hyun, I think if non-Apache Spark has some rules that should not be excluded, that should be added into non-Apache Spark specifically.

Also, this new configuration can still also be set by users similarly. It might be better to just document if there are somehow extra rules that should not be excluded.