@dongjoon-hyun 2019-04-11
@HyukjinKwon . This is about Apache Spark. If this is non-Apache issue, I don't need to create a PR here. What I mean by `custom` is `customOperatorOptimizationRules`.

1. Apache Spark provides `injectOptimizerRule` at `Extension`.
2. Apache Spark provides `set spark.sql.optimizer.excludedRules` to exclude any rule including the one injected by `Extension` at (1).

(1) and (2) is a good feature individually, but there is a conflict like this PR. There is no way to protect the injected optimizer rule. Please note that only (2) is dynamic. (1) and this PR is static configuration which is safe in SQL environment.