@HyukjinKwon 2019-04-12
But all the configurations are set by the similar way. Yes, static and dynamic difference is notable but still they can be set and unset.

Actually I saw a request about blacklisting some values for some configurations. How about we go in this way after a bit of discussion? IIRC (and as you know .. ), Livy supports a feature to blacklist a set of Spark configurations (https://github.com/cloudera/livy/blob/master/conf/spark-blacklist.conf.template).