@dongjoon-hyun 2019-04-12
Thanks! That sounds much nicer to me. +1 for the idea.