@dongjoon-hyun 2019-04-12
Do we have a JIRA issue and a PR for that? Or, shall I transform this to an initial PR for that?

If you want you can make a PR, it's also okay to me. Just let me know.