@HyukjinKwon 2019-04-12
I can take a look too (although it's going to take a bit of awhile ...). Let me know if you're busy.

I don't think there's a JIRA in Apache side because I met this needs with the previous company after you left .. :D

I am supportive of this approach at least for clarification. To me, I found this needs is reasonable but also not much clever workaround too.