@dongjoon-hyun 2019-04-12
Okay, then, @gatorsmile , do you agree with introducing a new general configuration file `blacklist.conf.template` instead of this specific configuration?