@dongjoon-hyun 2019-04-15
@beliefer . Thank you for review. For the following question, please see my previous comments https://github.com/apache/spark/pull/24339#issuecomment-482271610 and https://github.com/apache/spark/pull/24339#issuecomment-482071376 .
> I don't understand the admin who will use this feature. Why we need this change?