@rxin 2019-04-15
I agree with the above -- why have this? Pretty confusing ... Users / admins that use any of the super advanced features should know what they are doing, before attempting to do that. If one really wants to go out of their way to exclude specific rules, there might be a chance that they actually want to exclude a rule that's considered "should not exclude", especially if they end up writing their version of the rule.