@beliefer 2019-04-16
> @beliefer . Thank you for review. For the following question, please see my previous comments [#24339 (comment)](https://github.com/apache/spark/pull/24339#issuecomment-482271610) and [#24339 (comment)](https://github.com/apache/spark/pull/24339#issuecomment-482071376) .
> 
> > I don't understand the admin who will use this feature. Why we need this change?
> 
> Also, please see Apache Livy's blacklist feature advised in this PR, [#24339 (comment)](https://github.com/apache/spark/pull/24339#issuecomment-482414356) , too.

Yes, `spark.sql.optimizer.excludedRules` is dangerous, especially a server like Livy or Thrift.