@dongjoon-hyun 2019-04-16
Thank you for the decision, @rxin . I'll close this PR. Thank you, @gatorsmile , @HyukjinKwon , @MaxGekk , @beliefer , too.

cc @dbtsai and @beettlle This will not be a part of official Apache Spark.